test shouldRetryCancel on cancelled request and timing out requests + remove test that was only run in 1.5
